+++ octasine/src/gui/interface/mod.rs
@@ -150,7 +150,6 @@ impl<H: GuiSyncHandle> OctaSineIcedApplication<H> {
                     OperatorParameter::FrequencyFree => operator.frequency_free.set_value(v),
                     OperatorParameter::FrequencyFine => operator.frequency_fine.set_value(v),
                     OperatorParameter::AttackDuration => operator.envelope.set_attack_duration(v),
-                    OperatorParameter::AttackValue => operator.envelope.set_attack_end_value(v),
                     OperatorParameter::DecayDuration => operator.envelope.set_decay_duration(v),
                     OperatorParameter::DecayValue => operator.envelope.set_decay_end_value(v),
                     OperatorParameter::ReleaseDuration => operator.envelope.set_release_duration(v),